+++ app/src/main/java/com/aliayali/personalfinanceapp/domain/repository/NoteRepository.kt
@@ -1,6 +1,5 @@
 package com.aliayali.personalfinanceapp.domain.repository
 
-import androidx.room.Delete
 import com.aliayali.personalfinanceapp.data.local.database.entity.NoteEntity
 import kotlinx.coroutines.flow.Flow
 
@@ -9,6 +8,6 @@ interface NoteRepository {
     suspend fun updateNote(note: NoteEntity)
     suspend fun deleteNote(note: NoteEntity)
     fun getAllNotes(): Flow<List<NoteEntity>>
-    suspend fun getNoteById(noteId: Int): NoteEntity?
+    suspend fun getNoteById(noteId: Long): NoteEntity?
     fun getNotesByDate(date: String): Flow<List<NoteEntity>>
 }
